+++ tests/meta/orders/conftest.py
@@ -5,6 +5,7 @@ from tests.mocks.osticket import MockOsTicket
 from cg.meta.orders import OrdersAPI
 from cg.meta.orders.api import FastqSubmitter
 from cg.meta.orders.metagenome_submitter import MetagenomeSubmitter
+from cg.meta.orders.microbial_submitter import MicrobialSubmitter
 from cg.meta.orders.status import StatusHandler
 from cg.meta.orders.ticket_handler import TicketHandler
 from cg.models.orders.order import OrderIn, OrderType
@@ -71,7 +72,7 @@ def microbial_status_data(microbial_order_to_submit: dict):
     """Parse microbial order example."""
     project: OrderType = OrderType.MICROSALT
     order: OrderIn = OrderIn.parse_obj(obj=microbial_order_to_submit, project=project)
-    return StatusHandler.microbial_samples_to_status(order=order)
+    return MicrobialSubmitter.microbial_samples_to_status(order=order)
 
 
 @pytest.fixture
