+++ tests/meta/orders/test_meta_orders_api.py
@@ -6,14 +6,14 @@ from cg.exc import OrderError
 from cg.meta.orders import OrdersAPI
 from cg.models.orders.order import OrderIn, OrderType
 from cg.models.orders.samples import MipDnaSample
-from cg.store import models
+from cg.store import models, Store
 from cgmodels.cg.constants import Pipeline
 import cg
 from tests.store_helpers import StoreHelpers
 
 PROCESS_LIMS_FUNCTION_OLD = "cg.meta.orders.api.process_lims"
 PROCESS_LIMS_FUNCTION = "cg.meta.orders.lims.process_lims"
-SUBMITTERS = ["api", "lims", "fastq_submitter", "metagenome_submitter"]
+SUBMITTERS = ["api", "lims", "fastq_submitter", "metagenome_submitter", "microbial_submitter"]
 
 
 def test_too_long_order_name():
@@ -202,7 +202,7 @@ def test_submit_duplicate_sample_case_name(
     ticket_number: int,
     user_name: str,
     user_mail: str,
-    mocker,
+    monkeypatch,
 ):
     # GIVEN we have an order with a case that is already in the database
     order_data = OrderIn.parse_obj(obj=all_orders_to_submit[order_type], project=order_type)
@@ -222,8 +222,13 @@ def test_submit_duplicate_sample_case_name(
             store.add_commit(case_obj)
         assert store.find_family(customer=customer_obj, name=case_id)
 
-    mocker.patch(PROCESS_LIMS_FUNCTION_OLD)
-    mocker.patch(PROCESS_LIMS_FUNCTION)
+    lims_project_data = {"id": "ADM1234", "date": dt.datetime.now()}
+    lims_map = {sample.name: f"ELH123A{index}" for index, sample in enumerate(order_data.samples)}
+    for submitter in SUBMITTERS:
+        monkeypatch.setattr(
+            f"cg.meta.orders.{submitter}.process_lims",
+            lambda **kwargs: (lims_project_data, lims_map),
+        )
 
     # WHEN calling submit
     # THEN an OrderError should be raised on duplicate case name
@@ -231,9 +236,6 @@ def test_submit_duplicate_sample_case_name(
         orders_api.submit(
             project=order_type, order_in=order_data, user_name=user_name, user_mail=user_mail
         )
-
-    # Then no new samples should have been created in LIMS
-    cg.meta.orders.lims.process_lims.assert_not_called()
 
 
 def test_submit_unique_sample_case_name(
@@ -429,20 +431,25 @@ def test_submit_unique_sample_name(
     [OrderType.SARS_COV_2],
 )
 def test_sarscov2_submit_duplicate_sample_name(
-    orders_api,
-    sample_store,
+    orders_api: OrdersAPI,
+    sample_store: Store,
     all_orders_to_submit,
-    order_type,
+    order_type: OrderType,
     ticket_number: int,
     user_name: str,
     user_mail: str,
-    mocker,
+    monkeypatch,
     helpers,
 ):
     # GIVEN we have an order with samples that is already in the database
     order_data = OrderIn.parse_obj(obj=all_orders_to_submit[order_type], project=order_type)
-    mocker.patch(PROCESS_LIMS_FUNCTION_OLD)
-    mocker.patch(PROCESS_LIMS_FUNCTION)
+    lims_project_data = {"id": "ADM1234", "date": dt.datetime.now()}
+    lims_map = {sample.name: f"ELH123A{index}" for index, sample in enumerate(order_data.samples)}
+    for submitter in SUBMITTERS:
+        monkeypatch.setattr(
+            f"cg.meta.orders.{submitter}.process_lims",
+            lambda **kwargs: (lims_project_data, lims_map),
+        )
 
     store = orders_api.status
     customer_obj = store.customer(order_data.customer)
@@ -463,9 +470,6 @@ def test_sarscov2_submit_duplicate_sample_name(
             project=order_type, order_in=order_data, user_name=user_name, user_mail=user_mail
         )
 
-    # Then no new samples should have been created in LIMS
-    cg.meta.orders.lims.process_lims.assert_not_called()
-
 
 @pytest.mark.parametrize(
     "order_type",
@@ -480,15 +484,15 @@ def test_sarscov2_submit_duplicate_sample_name(
     ],
 )
 def test_not_sarscov2_submit_duplicate_sample_name(
-    orders_api,
-    sample_store,
+    orders_api: OrdersAPI,
+    sample_store: Store,
     all_orders_to_submit,
     order_type: OrderType,
     ticket_number: int,
     user_name: str,
     user_mail: str,
     monkeypatch,
-    helpers,
+    helpers: StoreHelpers,
 ):
     # GIVEN we have an order with samples that is already in the database
     order_data = OrderIn.parse_obj(obj=all_orders_to_submit[order_type], project=order_type)
